requirements.txt
together>=1.2.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
streamlit>=1.28.0
python-docx>=0.8.11
PyPDF2>=3.0.0
openpyxl>=3.1.0
Pillow>=10.0.0
pytesseract>=0.3.10
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.20
scikit-learn>=1.3.0
wordcloud>=1.9.0
